% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{data}
\name{Available classes}
\alias{Available classes}
\alias{CLASSES}
\title{Available classes.}
\format{\preformatted{ Named num [1:3] 1 2 999
 - attr(*, "names")= chr [1:3] "positive" "negative" "unknown"
}}
\usage{
CLASSES
}
\description{
Important!
This vector MUST BE synchronized with class values in PHP application!
}
\keyword{datasets}

