% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{data}
\name{Available metrics}
\alias{AVAILABLE_METRICS}
\alias{Available metrics}
\title{Fields definition}
\format{\preformatted{ Named chr [1:43] "factor" "character" "character" "factor" "character" "character" "character" ...
 - attr(*, "names")= chr [1:43] "class" "domain" "url" "http_code" ...
}}
\usage{
AVAILABLE_METRICS
}
\description{
Fields definition
}
\keyword{datasets}

