% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{classify_metrics}
\alias{classify_metrics}
\title{Classify link by metrics, for single link}
\usage{
classify_metrics(values, labels)
}
\arguments{
\item{values}{}

\item{label}{}
}
\value{
Classification info about metrics
}
\description{
Classify link by metrics, for single link
}

