% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{clean_raw_data}
\alias{clean_raw_data}
\title{Clean data frame from unused columns and NA's values.}
\usage{
clean_raw_data(data)
}
\arguments{
\item{data}{Data frame with metrics}
}
\value{
Clean data frame
}
\description{
Clean data frame from unused columns and NA's values.
}

