% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{load_classification}
\alias{load_classification}
\title{Load url classification data}
\usage{
load_classification(path)
}
\arguments{
\item{path}{}
}
\value{
Data frame with url classification
}
\description{
Load url classification data
}

