% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{prepare_metrics}
\alias{prepare_metrics}
\title{Prepare labeled metrics from file.}
\usage{
prepare_metrics(metricsPath, classificationPath)
}
\arguments{
\item{path}{}

\item{path}{}
}
\value{
Data frame with cleaned and labeled metrics
}
\description{
Prepare labeled metrics from file.
}

