% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{data}
\name{Supported metrics}
\alias{SUPPORTED_METRICS}
\alias{Supported metrics}
\title{Fields supported by model.}
\format{\preformatted{ chr [1:30] "class" "http_code" "domain_moz_domain_authority" "domain_moz_mozrank" ...
}}
\usage{
SUPPORTED_METRICS
}
\description{
Fields supported by model.
}
\keyword{datasets}

